refactor: 이미지 포맷 변경 (webp) 및 리사이징
