+++ frontend/src/pages/landing/LandingPage.tsx
@@ -9,13 +9,13 @@ import {
 import { useRef, useState } from 'react';
 import { MdArrowUpward, MdKeyboardDoubleArrowDown } from 'react-icons/md';
 import { useNavigate } from 'react-router-dom';
-import completeImage from '../../@assets/images/screenshots/complete.png';
-import guestbookImage from '../../@assets/images/screenshots/guestbook.png';
-import hostMainImage from '../../@assets/images/screenshots/host-main.png';
-import introduceImage from '../../@assets/images/screenshots/introduce.png';
-import spaceImage from '../../@assets/images/screenshots/space.png';
-import visibilityImage from '../../@assets/images/screenshots/visibility.png';
-import titleImage from '../../@assets/logo/main-logo.png';
+import completeImage from '../../@assets/images/screenshots/complete.png?w=600&format=webp&srcset';
+import guestbookImage from '../../@assets/images/screenshots/guestbook.png?w=600&format=webp&srcset';
+import hostMainImage from '../../@assets/images/screenshots/host-main.png?w=600&format=webp&srcset';
+import introduceImage from '../../@assets/images/screenshots/introduce.png?w=600&format=webp&srcset';
+import spaceImage from '../../@assets/images/screenshots/space.png?w=600&format=webp&srcset';
+import visibilityImage from '../../@assets/images/screenshots/visibility.png?w=600&format=webp&srcset';
+import titleImage from '../../@assets/logo/main-logo.png?w=600&format=webp&srcset';
 import Button from '../../components/@common/buttons/button/Button';
 import IconButton from '../../components/@common/buttons/iconButton/IconButton';
 import { AUTH_COOKIES } from '../../constants/cookie';
@@ -142,6 +142,8 @@ const LandingPage = () => {
     });
     navigate(ROUTES.AUTH.LOGIN);
   };
+
+  console.log('here', titleImage, titleImage.src, titleImage.srcset);
 
   return (
     <S.Wrapper>
