+++ frontend/src/types/globals/imagetools.d.ts
@@ -1,0 +1,6 @@
+declare module '*&imagetools';
+declare module '*?imagetools';
+
+declare module '*?format=webp';
+declare module '*?w=*';
+declare module '*&srcset';
